split amazon test case into separate source file
